Reworked arrays.
Converted address array to the generic class instead.

Note:
Some people are confused by the fact u64 can be used to arrays since it's actually an array inside an array but it's whatever. Ill just reduce the amount of possibilities to u32.MAX_VALUE - 1 and call it a day. It's now all u32.

This will resolve people confusions and nobody will do mistakes.